requirements.txt
flask==2.0.1
werkzeug==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.1
pg8000==1.29.0